Remove GetValidHostAliases from config setup
